Add initial support for running via an on-system JIT
This commit adds experimental support (aka off-by-default) to execute
wasm files in a precompiled release mode JIT exposed as a dynamic
library through the proposed wasm C API [1]. A build script determines
what's used, and for now it's only wired up to work on my own system. I
figured it'd be good to get this into CI though to test out and ensure
others can tinker around with this as well.

[1]: https://github.com/WebAssembly/wasm-c-api