Start prototyping bulk-transfers of data
This commit switches to per-API function calls into the host runtime to
instead minimizing the amount of times a wasm blob calls the host
runtime. This requires a bulk transfer of the entire `TokenStream` up
front into the wasm blob, along with a bulk transfer from wasm back into
the native runtime afterwards.

Current a bincode-like encoding is used to serialize and deserialize
`TokenStream`, with the data format duplicated across a few
`{encode,decode}.rs` files. Eventually it'd probably be good to use a
macro_rules macro or similar to avoid this duplication, but for now it's
not a whole lot of code anyway.

This makes the wasm proc macro sort of hard for now, but I hope to
improve that in subsequent commits.

Additionally the `proc-macro2` shadow crate has been updated to more
closely match the fallback of `proc-macro2` itself, storing most
information in-memory as opposed to storing handles.