Improve JIT build documentation and support
Add a section to the README about how to use `WATT_JIT` as well as an
example of how to build wasmtime and use it. This also in the process
switches from static linkage to dynamic linkage discovered at runtime to
avoid having to deal with things like `LD_LIBRARY_PATH` with proc macros
(dylibs) depending on dylibs (wasmtime).