Tweak README example to be clearer.
Both the crate and the macro in the original text were named "my_macro",
so it was not obvious which one corresponded to what. This commit names them
"foo-macros" and "bar".

Also, the original text talked about the "my_macros" crate and the shim crate,
but the shim crate deserves to have the name the user gave to the original
pre-watt crate, since that is what gets published to crates.io.
The text now walks the user through renaming "foo-macros" to "foo-macros-impl"
and naming the shim crate "foo-macros".