Add Display impl to LexError
Starting with syn 2.0.47, syn reuses the LexError Diplay trait for its own
error. This was not implemented for the patched proc-macro2 type, leading to
errors such as:

  error[E0277]: `LexError` doesn't implement `std::fmt::Display`
     --> /home/mwanzenboeck/.cargo/registry/src/index.crates.io-6f17d22bba15001f/syn-2.0.86/src/error.rs:407:32
      |
  407 |         Error::new(err.span(), err)
      |         ----------             ^^^ `LexError` cannot be formatted with the default formatter
      |         |
      |         required by a bound introduced by this call
      |
      = help: the trait `std::fmt::Display` is not implemented for `LexError`
      = note: in format strings you may be able to use `{:?}` (or {:#?} for pretty-print) instead

To fix this, implement Display for LexError, using the fallback error message,
since we do not store any details on the exact error.